Converted the project to install the OPC Foundation jar in the local maven repository to avoid the practice of putting local directory structures on the build. Updated the Readme.md to reflect the build process. Commented out generic or empty tests.
