+++ nifi-opc-ua/nifi-opcua-bundle/nifi-opcua-service/src/main/java/com/kentender/nifi/nifi_opcua_services/Utils.java
