+++ nifi-opc-ua/nifi-opcua-bundle/nifi-opcua-service/src/test/java/com/kentender/nifi/nifi_opcua_services/TestStandardMyService.java
@@ -14,26 +14,31 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.kentender.nifi.nifi_opcua_bundle;
+package com.kentender.nifi.nifi_opcua_services;
 
+import org.apache.nifi.reporting.InitializationException;
 import org.apache.nifi.util.TestRunner;
 import org.apache.nifi.util.TestRunners;
 import org.junit.Before;
 import org.junit.Test;
 
-
-public class MyProcessorTest {
-
-    private TestRunner testRunner;
+public class TestStandardMyService {
 
     @Before
     public void init() {
-        testRunner = TestRunners.newTestRunner(MyProcessor.class);
-    }
-
-    @Test
-    public void testProcessor() {
 
     }
+
+/*    @Test
+    public void testService() throws InitializationException {
+        final TestRunner runner = TestRunners.newTestRunner(TestProcessor.class);
+        final StandardMyService service = new StandardMyService();
+        runner.addControllerService("test-good", service);
+
+        runner.setProperty(service, StandardMyService.MY_PROPERTY, "test-value");
+        runner.enableControllerService(service);
+
+        runner.assertValid(service);
+    }*/
 
 }
