+++ nifi-opc-ua/nifi-opcua-bundle/pom.xml
@@ -1,0 +1,43 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+  Licensed to the Apache Software Foundation (ASF) under one or more
+  contributor license agreements. See the NOTICE file distributed with
+  this work for additional information regarding copyright ownership.
+  The ASF licenses this file to You under the Apache License, Version 2.0
+  (the "License"); you may not use this file except in compliance with
+  the License. You may obtain a copy of the License at
+  http://www.apache.org/licenses/LICENSE-2.0
+  Unless required by applicable law or agreed to in writing, software
+  distributed under the License is distributed on an "AS IS" BASIS,
+  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  See the License for the specific language governing permissions and
+  limitations under the License.
+-->
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <groupId>com.kentender.nifi</groupId>
+    <artifactId>nifi-opcua-bundle</artifactId>
+    <version>0.0.1-SNAPSHOT</version>
+    <packaging>pom</packaging>
+    <parent>
+        <artifactId>nifi-opcua</artifactId>
+        <groupId>com.kentender.nifi</groupId>
+        <version>0.0.1-SNAPSHOT</version>
+    </parent>
+    <modules>
+        <module>nifi-opcua-bundle-processors</module>
+        <module>nifi-opcua-bundle-nar</module>
+        <module>nifi-opcua-service</module>
+        <module>nifi-opcua-service-api</module>
+        <module>nifi-opcua-service-api-nar</module>
+        <module>nifi-opcua-service-nar</module>
+    </modules>
+
+    <dependencies>
+        <dependency>
+            <groupId>org.opcfoundation.ua</groupId>
+            <artifactId>opc-ua-stack</artifactId>
+            <version>${opc-ua-stack.version}</version>
+        </dependency>
+    </dependencies>
+</project>
