+++ nifi-opc-ua/pom.xml
@@ -15,34 +15,24 @@
 -->
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-
+    <groupId>com.kentender.nifi</groupId>
+    <artifactId>nifi-opcua</artifactId>
+    <version>0.0.1-SNAPSHOT</version>
+    <packaging>pom</packaging>
     <parent>
         <groupId>org.apache.nifi</groupId>
         <artifactId>nifi-nar-bundles</artifactId>
         <version>1.1.0</version>
     </parent>
-
-    <groupId>com.kentender.nifi</groupId>
-    <artifactId>nifi-opcua-bundle</artifactId>
-    <version>0.0.1-SNAPSHOT</version>
-    <packaging>pom</packaging>
+    <properties>
+        <jdk.version>1.8</jdk.version>
+        <opc-ua-stack.version>1.03.341.0-SNAPSHOT</opc-ua-stack.version>
+    </properties>
 
     <modules>
-        <module>nifi-opcua-bundle-processors</module>
-        <module>nifi-opcua-bundle-nar</module>
-        <module>nifi-opcua-service</module>
-        <module>nifi-opcua-service-api</module>
-        <module>nifi-opcua-service-api-nar</module>
-        <module>nifi-opcua-service-nar</module>
+        <module>opc-deploy-local</module>
+        <module>nifi-opcua-bundle</module>
     </modules>
 
-    <dependencies>
-    	<dependency>
-    		<groupId>org.opcfoundation.ua</groupId>
-    		<artifactId>opc-ua-stack</artifactId>
-    		<version>1.03.340.0-SNAPSHOT</version>
-    		<scope>system</scope>
-    		<systemPath>C:\Users\wades\git\UA-Java\target\opc-ua-stack-1.03.340.0-SNAPSHOT.jar</systemPath>
-    	</dependency>
-    </dependencies>
+
 </project>
